Call to_dict method instead of __dict__
This is because __dict__ gives the exact naming of the class fields
(e.g., _study_table_name). Notice that the underscore at the first index
is included. By utilising the to_dict method, it returns the correct
field names so that can be utilised properly in the client admin
program.