Validate input on start screen to avoid invalid user emails
This is done using HTML and JavaScript. First, the email input is
checked to see whether they are actual emails or not using type = email
on the html input tag. Second, JavaScript is used to make sure that the
email and confirmation email are the same. If the 2 of these are not
satisfied yet, the user can't go to the next screen, which is the
booking screen