Enable mobile os detection to enable generic date detection
This change is made since iOS (Apple) doesn't work really well with
usual date format (i.e. YYYY-MM-DD), it needs the format (MM/DD/YYYY),
hence detecting what OS it is is important. If it's android, we'll work
with the first format, if it's apple, we'll work with the second format.